feat(instant-start): move instant start export to starter download (#237)
