+++ src/lib/command.ts
@@ -167,7 +167,6 @@ type CreateCommandRouterConfig = {
 	name: string;
 	description: string;
 	sections?: Record<string, string>;
-	defaultHandler?: () => Promise<void>;
 	commands: Record<string, RouterCommand>;
 };
 type RouterCommand = { handler: () => Promise<void>; description: string };
@@ -177,11 +176,6 @@ export function createCommandRouter(config: CreateCommandRouterConfig): () => Pr
 
 	return async function () {
 		const args = process.argv.slice(1 + depth);
-
-		if (config.defaultHandler && (args.length === 0 || args[0].startsWith("-"))) {
-			await config.defaultHandler();
-			return;
-		}
 
 		const {
 			positionals: [subcommand],
