+++ src/lib/zip.ts
@@ -3,7 +3,7 @@ import { mkdir, mkdtemp, readdir, rename, rm, stat, writeFile } from "node:fs/pr
 import { basename, dirname, isAbsolute, join, win32 } from "node:path";
 
 export async function extractZip(data: Uint8Array, destination: string): Promise<void> {
-	await assertEmptyOrMissingDirectory(destination);
+	const destinationExisted = await assertEmptyOrMissingDirectory(destination);
 
 	const parent = dirname(destination);
 	await mkdir(parent, { recursive: true });
@@ -28,11 +28,14 @@ export async function extractZip(data: Uint8Array, destination: string): Promise
 		await rename(temporaryDirectory, destination);
 	} catch (error) {
 		await rm(temporaryDirectory, { recursive: true, force: true });
+		if (destinationExisted) {
+			await mkdir(destination, { recursive: true });
+		}
 		throw error;
 	}
 }
 
-async function assertEmptyOrMissingDirectory(destination: string): Promise<void> {
+async function assertEmptyOrMissingDirectory(destination: string): Promise<boolean> {
 	try {
 		const destinationStat = await stat(destination);
 		if (!destinationStat.isDirectory()) {
@@ -41,8 +44,9 @@ async function assertEmptyOrMissingDirectory(destination: string): Promise<void>
 		if ((await readdir(destination)).length > 0) {
 			throw new Error(`Destination directory is not empty: ${destination}`);
 		}
+		return true;
 	} catch (error) {
-		if (isMissingFileError(error)) return;
+		if (isMissingFileError(error)) return false;
 		throw error;
 	}
 }
