+++ test/starter-download.test.ts
@@ -2,42 +2,51 @@ import { zipSync } from "fflate";
 import { access, mkdir, mkdtemp, readFile, rm, writeFile } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
+import { pathToFileURL } from "node:url";
 import { afterEach, describe, expect, it as unitTest, onTestFinished, vi } from "vitest";
 
-import { removePreviewsByURL } from "../src/lib/prismic/clients/core";
+import { findPackageJson } from "../src/lib/packageJson";
+import { addPreview, hasPreviewByURL, removePreviewsByURL } from "../src/lib/prismic/clients/core";
 import { getOrCreateInstantStartExport } from "../src/lib/prismic/clients/website-generator";
 import { extractZip } from "../src/lib/zip";
 import { captureOutput, it } from "./it";
 
-it("supports init instant --help", async ({ expect, prismic }) => {
-	const { stdout, exitCode } = await prismic("init", ["instant", "--help"]);
+it("supports starter --help", async ({ expect, prismic }) => {
+	const { stdout, exitCode } = await prismic("starter", ["--help"]);
 	expect(exitCode).toBe(0);
-	expect(stdout).toContain("prismic init instant [options]");
+	expect(stdout).toContain("prismic starter <command>");
+	expect(stdout).toContain("download");
+});
+
+it("supports starter download --help", async ({ expect, prismic }) => {
+	const { stdout, exitCode } = await prismic("starter", ["download", "--help"]);
+	expect(exitCode).toBe(0);
+	expect(stdout).toContain("prismic starter download [options]");
 	expect(stdout).toContain("--repo string");
 	expect(stdout).toContain("(required)");
 });
 
-it("requires --repo in instant mode", async ({ expect, prismic }) => {
-	const { stderr, exitCode } = await prismic("init", ["instant"]);
+it("requires --repo", async ({ expect, prismic }) => {
+	const { stderr, exitCode } = await prismic("starter", ["download"]);
 	expect(exitCode).toBe(1);
 	expect(stderr).toContain("Missing required option: --repo");
 });
 
-it("rejects an unknown init subcommand", async ({ expect, prismic }) => {
-	const { stderr, exitCode } = await prismic("init", ["unknown"]);
+it("rejects an unknown starter subcommand", async ({ expect, prismic }) => {
+	const { stderr, exitCode } = await prismic("starter", ["unknown"]);
 	expect(exitCode).toBe(1);
 	expect(stderr).toContain("Unknown command: unknown");
 });
 
-it("rejects extra instant arguments", async ({ expect, prismic }) => {
-	const { stderr, exitCode } = await prismic("init", ["instant", "extra", "--repo", "my-repo"]);
+it("rejects extra download arguments", async ({ expect, prismic }) => {
+	const { stderr, exitCode } = await prismic("starter", ["download", "extra", "--repo", "my-repo"]);
 	expect(exitCode).toBe(1);
 	expect(stderr).toContain("extra");
 });
 
-it("rejects --lang in instant mode", async ({ expect, prismic }) => {
-	const { stderr, exitCode } = await prismic("init", [
-		"instant",
+it("rejects init-only options", async ({ expect, prismic }) => {
+	const { stderr, exitCode } = await prismic("starter", [
+		"download",
 		"--repo",
 		"my-repo",
 		"--lang",
@@ -47,40 +56,53 @@ it("rejects --lang in instant mode", async ({ expect, prismic }) => {
 	expect(stderr).toContain("--lang");
 });
 
-it("rejects --no-setup in instant mode", async ({ expect, prismic }) => {
-	const { stderr, exitCode } = await prismic("init", [
-		"instant",
+it("validates the repository name", async ({ expect, prismic }) => {
+	const { stderr, exitCode } = await prismic("starter", [
+		"download",
 		"--repo",
-		"my-repo",
-		"--no-setup",
+		"invalid_repository",
 	]);
 	expect(exitCode).toBe(1);
-	expect(stderr).toContain("--no-setup");
-});
-
-it("dispatches instant mode before checking the current project", async ({ expect, prismic }) => {
-	const { stderr, exitCode } = await prismic("init", ["instant", "--repo", "invalid_repository"]);
-	expect(exitCode).toBe(1);
 	expect(stderr).toContain("Invalid repository name");
-	expect(stderr).not.toContain("already initialized");
 });
 
-it("uses the init login flow", async ({ expect, logout, prismic, repo }) => {
+it("validates repository access", async ({ expect, login, prismic }) => {
+	await login();
+	const { stderr, exitCode } = await prismic("starter", [
+		"download",
+		"--repo",
+		"missing-repository",
+	]);
+	expect(exitCode).toBe(1);
+	expect(stderr).toContain('Repository "missing-repository" not found in your account');
+});
+
+it("requires Type Builder", async ({ expect, login, prismic, repo }) => {
+	await login();
+	const { stderr, exitCode } = await prismic("starter", ["download", "--repo", repo], {
+		nodeOptions: { env: { PRISMIC_TYPE_BUILDER_ENABLED: "false" } },
+	});
+	expect(exitCode).toBe(1);
+	expect(stderr).toContain("This command requires the Type Builder");
+});
+
+it("uses the browser login flow", async ({ expect, logout, prismic, repo }) => {
 	await logout();
-	const proc = prismic("init", ["instant", "--repo", repo, "--no-browser"]);
+	const proc = prismic("starter", ["download", "--repo", repo, "--no-browser"]);
 	const output = captureOutput(proc);
 
 	await expect.poll(output, { timeout: 15_000 }).toMatch(/port=(\d+)/);
 	proc.kill();
 });
 
-it("does not list instant-start as a top-level command", async ({ expect, prismic }) => {
+it("lists starter as a top-level command", async ({ expect, prismic }) => {
 	const { stdout, exitCode } = await prismic("", ["--help"]);
 	expect(exitCode).toBe(0);
+	expect(stdout).toContain("starter");
 	expect(stdout).not.toContain("instant-start");
 });
 
-describe.sequential("Instant Start API client", () => {
+describe.sequential("Starter download API client", () => {
 	afterEach(() => {
 		vi.unstubAllGlobals();
 	});
@@ -165,6 +187,73 @@ describe.sequential("Instant Start API client", () => {
 		expect(url.toString()).toContain("/previews/delete/starter-preview");
 		expect(init?.method).toBe("POST");
 	});
+
+	unitTest("adds the local development preview", async () => {
+		const fetchMock = vi.fn<typeof fetch>().mockResolvedValue(new Response(null, { status: 200 }));
+		vi.stubGlobal("fetch", fetchMock);
+
+		await addPreview(
+			{
+				name: "Development",
+				websiteURL: "http://localhost:3000",
+				resolverPath: "/api/preview",
+			},
+			{
+				repo: "my-repo",
+				token: "test-token",
+				host: "prismic.io",
+			},
+		);
+
+		expect(fetchMock).toHaveBeenCalledOnce();
+		const [url, init] = fetchMock.mock.calls[0];
+		expect(url.toString()).toContain("/previews/new");
+		expect(init?.method).toBe("POST");
+		expect(init?.body).toBe(
+			JSON.stringify({
+				name: "Development",
+				websiteURL: "http://localhost:3000",
+				resolverPath: "/api/preview",
+			}),
+		);
+	});
+
+	unitTest("finds an existing local development preview", async () => {
+		const fetchMock = vi.fn<typeof fetch>().mockResolvedValue(
+			jsonResponse({
+				results: [
+					{
+						id: "development-preview",
+						label: "Development",
+						url: "http://localhost:3000/api/preview",
+					},
+				],
+			}),
+		);
+		vi.stubGlobal("fetch", fetchMock);
+
+		await expect(
+			hasPreviewByURL("http://localhost:3000/api/preview", {
+				repo: "my-repo",
+				token: "test-token",
+				host: "prismic.io",
+			}),
+		).resolves.toBe(true);
+	});
+});
+
+describe("Package installation", () => {
+	unitTest("does not find package.json above the extracted project", async () => {
+		const root = await makeTemporaryDirectory();
+		const destination = join(root, "my-repo");
+		await mkdir(destination);
+		await writeFile(join(root, "package.json"), "{}");
+		const destinationUrl = pathToFileURL(`${destination}/`);
+
+		await expect(findPackageJson({ start: destinationUrl, stop: destinationUrl })).rejects.toThrow(
+			"Could not find a package.json file",
+		);
+	});
 });
 
 describe("ZIP extraction", () => {
@@ -227,7 +316,7 @@ function jsonResponse(value: unknown): Response {
 }
 
 async function makeTemporaryDirectory(): Promise<string> {
-	const directory = await mkdtemp(join(tmpdir(), "prismic-instant-start-"));
+	const directory = await mkdtemp(join(tmpdir(), "prismic-starter-download-"));
 	onTestFinished(() => rm(directory, { recursive: true, force: true }));
 	return directory;
 }
